Stronger ToolProvidedMetadata abstraction.
- Move to its own file (goal is to be importable from set_metadata without baggage from tools and jobs modules).
- Proper interface with some basic documentation.
