+++ lib/galaxy/tools/parameters/output_collect.py
@@ -1,6 +1,5 @@
 """ Code allowing tools to define extra files associated with an output datset.
 """
-import json
 import logging
 import operator
 import os
@@ -22,127 +21,6 @@ from galaxy.util import (
 DATASET_ID_TOKEN = "DATASET_ID"
 
 log = logging.getLogger(__name__)
-
-
-class NullToolProvidedMetadata(object):
-
-    def get_new_datasets(self, output_name):
-        return []
-
-    def get_new_dataset_meta_by_basename(self, output_name, basename):
-        return {}
-
-    def has_failed_outputs(self):
-        return False
-
-    def get_unnamed_outputs(self):
-        return []
-
-
-class LegacyToolProvidedMetadata(object):
-
-    def __init__(self, job_wrapper, meta_file):
-        self.job_wrapper = job_wrapper
-        self.tool_provided_job_metadata = []
-
-        with open(meta_file, 'r') as f:
-            for line in f:
-                try:
-                    line = json.loads(line)
-                    assert 'type' in line
-                except Exception:
-                    log.exception('(%s) Got JSON data from tool, but data is improperly formatted or no "type" key in data' % job_wrapper.job_id)
-                    log.debug('Offending data was: %s' % line)
-                    continue
-                # Set the dataset id if it's a dataset entry and isn't set.
-                # This isn't insecure.  We loop the job's output datasets in
-                # the finish method, so if a tool writes out metadata for a
-                # dataset id that it doesn't own, it'll just be ignored.
-                if line['type'] == 'dataset' and 'dataset_id' not in line:
-                    try:
-                        line['dataset_id'] = job_wrapper.get_output_file_id(line['dataset'])
-                    except KeyError:
-                        log.warning('(%s) Tool provided job dataset-specific metadata without specifying a dataset' % job_wrapper.job_id)
-                        continue
-                self.tool_provided_job_metadata.append(line)
-
-    def get_meta_by_dataset_id(self, dataset_id):
-        for meta in self.tool_provided_job_metadata:
-            if meta['type'] == 'dataset' and meta['dataset_id'] == dataset_id:
-                return meta
-
-    def get_new_dataset_meta_by_basename(self, output_name, basename):
-        for meta in self.tool_provided_job_metadata:
-            if meta['type'] == 'new_primary_dataset' and meta['filename'] == basename:
-                return meta
-
-    def get_new_datasets(self, output_name):
-        log.warning("Called get_new_datasets with legacy tool metadata provider - that is unimplemented.")
-        return []
-
-    def has_failed_outputs(self):
-        found_failed = False
-        for meta in self.tool_provided_job_metadata:
-            if meta.get("failed", False):
-                found_failed = True
-
-        return found_failed
-
-    def get_unnamed_outputs(self):
-        return []
-
-
-class ToolProvidedMetadata(object):
-
-    def __init__(self, job_wrapper, meta_file):
-        self.job_wrapper = job_wrapper
-        with open(meta_file, 'r') as f:
-            self.tool_provided_job_metadata = json.load(f)
-
-    def get_meta_by_name(self, name):
-        return self.tool_provided_job_metadata.get(name, {})
-
-    def get_new_dataset_meta_by_basename(self, output_name, basename):
-        datasets = self.tool_provided_job_metadata.get(output_name, {}).get("datasets", [])
-        for meta in datasets:
-            if meta['filename'] == basename:
-                return meta
-
-    def get_new_datasets(self, output_name):
-        datasets = self.tool_provided_job_metadata.get(output_name, {}).get("datasets", [])
-        if not datasets:
-            elements = self.tool_provided_job_metadata.get(output_name, {}).get("elements", [])
-            if elements:
-                datasets = self._elements_to_datasets(elements)
-        return datasets
-
-    def _elements_to_datasets(self, elements, level=0):
-        for element in elements:
-            extra_kwds = {"identifier_%d" % level: element["name"]}
-            if "elements" in element:
-                for inner_element in self._elements_to_datasets(element["elements"], level=level + 1):
-                    dataset = extra_kwds.copy()
-                    dataset.update(inner_element)
-                    yield dataset
-            else:
-                dataset = extra_kwds
-                extra_kwds.update(element)
-                yield extra_kwds
-
-    def has_failed_outputs(self):
-        found_failed = False
-        for output_name, meta in self.tool_provided_job_metadata.items():
-            if output_name == "__unnamed_outputs":
-                continue
-
-            if meta.get("failed", False):
-                found_failed = True
-
-        return found_failed
-
-    def get_unnamed_outputs(self):
-        log.debug("unnamed outputs [%s]" % self.tool_provided_job_metadata)
-        return self.tool_provided_job_metadata.get("__unnamed_outputs", [])
 
 
 def collect_dynamic_outputs(
